requirements.txt
torch==2.4.1
transformers>=4.42.0
accelerate>=0.33.0
peft>=0.11.0
bitsandbytes>=0.43.1
trl>=0.9.0
datasets
gradio
mlflow
sentencepiece
protobuf>=3.20.2,<6.0.0.dev0
python-dotenv
wandb  # optional for logging